requirements.txt
python-docx>=1.1.0
tkinterdnd2>=0.3.0
docxcompose>=1.3.6
comtypes>=1.1.14; sys_platform == 'win32'
